+++ src/inputscope/webui.py
@@ -4,9 +4,14 @@ Web frontend interface, displays statistics from a database.
 
 --quiet      prints out nothing
 
+------------------------------------------------------------------------------
+This file is part of InputScope - mouse and keyboard input visualizer.
+Released under the MIT License.
+
 @author      Erki Suurjaak
 @created     06.04.2015
-@modified    30.07.2023
+@modified    15.04.2024
+------------------------------------------------------------------------------
 """
 import collections
 import datetime
@@ -153,8 +158,8 @@ def inputdetail(input, table, period=None, session=None, appids=None, appnames=N
             if mycount >= conf.MaxEventsForStats: break # for myday
         if len(mydays) != len(days):
             where += [("day", ("IN", mydays))]
-    elif period and len(period) < 8: # Month period, query by known month days
-        mydays = [v["day"] for v in days if v["day"][:7] == period]
+    elif period and len(period) < 8: # Month/year period, query by known period days
+        mydays = [v["day"] for v in days if v["day"][:len(period)] == period]
         where += [("day", ("IN", mydays))]
     elif period:
         where += [("day", period)]
@@ -163,7 +168,7 @@ def inputdetail(input, table, period=None, session=None, appids=None, appnames=N
         count = db.fetchone(table, "COUNT(*) AS count", where=where)["count"]
     events = db.select(table, where=where, order="stamp", limit=conf.MaxEventsForStats)
     if "mouse" == input:
-        stats_texts, app_stats, heatmap_stats, events = stats_mouse(events, table, count)
+        stats_texts, app_stats, heatmap_sizes, heatmap_stats, events = stats_mouse(events, table, count)
     else:
         stats_texts, app_stats, events = stats_keyboard(events, table, count)
         heatmap_stats = db.fetch(table, "realkey AS key, COUNT(*) AS count", where, "realkey", "count DESC")
@@ -191,13 +196,19 @@ def inputindex(input):
     countminmax = "SUM(count) AS count, MIN(day) AS first, MAX(day) AS last"
     for table in conf.InputEvents[input]:
         stats[table] = db.fetchone("counts", countminmax, type=table)
-        periods, month = [], None
+        periods, month, year, months = [], None, None, 0
         for data in db.fetch("counts", "day AS period, count, 'day' AS class", order="day DESC", type=table):
+            if not month or month["period"][:4] != data["period"][:4]:
+                year = {"class": "year", "period": data["period"][:4], "count": 0}
+                periods.append(year)
             if not month or month["period"][:7] != data["period"][:7]:
                 month = {"class": "month", "period": data["period"][:7], "count": 0}
                 periods.append(month)
+                months += 1
             month["count"] += data["count"]
+            year["count"] += data["count"]
             periods.append(data)
+        if months < 2: periods = [x for x in periods if "year" != x["class"]]
         stats[table]["periods"] = periods
     dbinfo, sessions = stats_db(conf.DbPath), stats_sessions(input=input)
     return bottle.template("input.tpl", locals(), conf=conf)
@@ -280,7 +291,7 @@ def stats_keyboard(events, table, count):
     stats += [("Total unique %s" % table, len(uniques))]
     if deltas:
         stats += [("Total time interval", format_timedelta(last["dt"] - first["dt"]))]
-    app_items = [{"id": k, "cols": {"top": [a for a, _ in v.most_common(conf.MaxTopKeysForPrograms)]},
+    app_items = [{"id": k, "cols": {"top": [a for a, _ in v.most_common(conf.KeyboardTopForPrograms)]},
                   "path": appmap.get(k), "total": sum(v.values())} for k, v in app_stats.items()]
     app_results = collections.OrderedDict(
         (x["id"], x) for x in sorted(app_items, key=lambda x: x["total"], reverse=True)
@@ -298,18 +309,20 @@ def stats_mouse(events, table, count):
     first, last, totaldelta = None, None, datetime.timedelta()
     all_events = []
     HS = conf.MouseHeatmapSize
-    SZ = dict((k + 2, conf.DefaultScreenSize[k] / float(HS[k])) for k in [0, 1])
-    SZ.update({"dt": datetime.datetime.min, 0: 0, 1: 0}) # {0,1,2,3: x,y,w,h}
+    SZ = {0: 0, 1: 0, "dt": datetime.datetime.min} # {0,1,2,3,dt: xmin,ymin,w,h,startdt}
+    SZ.update((k + 2, conf.DefaultScreenSize[k]) for k in [0, 1])
     displayxymap  = collections.defaultdict(lambda: collections.defaultdict(int))
     counts, distances = collections.Counter(), collections.defaultdict(float)
     app_deltas = collections.defaultdict(datetime.timedelta) # {id: time in app}
     app_distances = collections.defaultdict(float) # {id: pixels}
-    SIZES = {} # Scale by desktop size at event time; {display: [{size}, ]}
+    SIZES = {} # Desktop sizes for event times, as {display: [{0,1,2,3,dt}, ]}
+    vals = lambda d, kk="xywh": [d[k] for k in kk]
     for row in db.fetch("screen_sizes", order="dt"):
-        row.update({0: row["x"], 2: row["w"] / float(HS[0]),
-                    1: row["y"], 3: row["h"] / float(HS[1])})
-        SIZES.setdefault(row["display"], []).append(row)
-    cursizes = {k: None for k in SIZES}
+        row.update({0: row["x"], 1: row["y"], 2: row["w"], 3: row["h"]})
+        if row["display"] not in SIZES or vals(SIZES[row["display"]][-1]) != vals(row):
+            SIZES.setdefault(row["display"], []).append(row)
+    cursizes = {k: None for k in SIZES} # {display: {0,1,2,3,dt}}
+    heatmap_sizes = {} # {display: (w, h)} scaled to screen size of first event
     for e in events:
         e["dt"] = datetime.datetime.fromtimestamp(e.pop("stamp"))
         if not first: first = e
@@ -328,11 +341,14 @@ def stats_mouse(events, table, count):
             # Find latest size from before event, fallback to first size recorded
             sz = next((s for s in sizes[::-1] if e["dt"] >= s["dt"]), sizes[0])
             cursizes[e["display"]] = sz
+        if e["display"] not in heatmap_sizes: # Make heatmap scaled to screen height
+            heatmap_sizes[e["display"]] = (HS[0], HS[0] * sz["h"] / sz["w"])
+        hs = heatmap_sizes[e["display"]]
 
         # Make heatmap coordinates, scaling event to screen size at event time
-        e["x"], e["y"] = [int((e["xy"[k]] - sz[k]) / sz[k + 2]) for k in [0, 1]]
-        # Constraint within heatmap, events at edges can have off-screen coordinates
-        e["x"], e["y"] = [max(0, min(e["xy"[k]], HS[k])) for k in [0, 1]]
+        e["x"], e["y"] = [int(float(e["xy"[k]] - sz[k]) * hs[k] / sz[k + 2]) for k in [0, 1]]
+        # Constrain within heatmap, events at edges can have off-screen coordinates
+        e["x"], e["y"] = [max(0, min(e["xy"[k]], hs[k])) for k in [0, 1]]
         displayxymap[e["display"]][(e["x"], e["y"])] += 1
         if "moves" == table:
             if appmap: app_stats[app_id].update([table])
@@ -395,7 +411,7 @@ def stats_mouse(events, table, count):
     app_results = collections.OrderedDict(
         (x["id"], x) for x in sorted(app_items, key=lambda x: x["total"], reverse=True)
     )
-    return stats, app_results, positions, all_events
+    return stats, app_results, heatmap_sizes, positions, all_events
 
 
 def stats_sessions(input=None):
@@ -429,7 +445,7 @@ def stats_db(filename):
     result = [("Database", filename),
               ("Created", datetime.datetime.fromtimestamp(os.path.getctime(filename))),
               ("Last modified", datetime.datetime.fromtimestamp(os.path.getmtime(filename))),
-              ("Size", format_bytes(os.path.getsize(filename))), ]
+              ("Size", format_bytes(db.get_size(filename))), ]
     counts = db.fetch("counts", "type, SUM(count) AS count", group="type")
     cmap = dict((x["type"], x["count"]) for x in counts)
     for name, tables in conf.InputTables:
@@ -437,6 +453,7 @@ def stats_db(filename):
         result += [("%s events" % name.capitalize(), countstr)]
     result += [("Sessions", db.fetchone("sessions", "COUNT(*) AS count")["count"])]
     result += [("%s version" % conf.Title, "%s (%s)" % (conf.Version, conf.VersionDate))]
+    result += [("Configuration", conf.ConfigPath or "")]
     return result